requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
supabase==2.0.0
postgrest==0.13.0
psycopg2-binary==2.9.9

# HTTP clients
httpx==0.25.2
redis==5.0.1

# Docker
docker==6.1.3

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0

# Billing
stripe==7.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0

